// web/src/api/generated/models/salesDraftItemResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * POS Cafe API
 * Backend API for Cafe Point of Sale System built with Vertical Slice Architecture and Idiomatic Go.
 * OpenAPI spec version: 1.0
 */
import type { SalesSelectedModifierOptionResponse } from './salesSelectedModifierOptionResponse';

export interface SalesDraftItemResponse {
  available?: boolean;
  id?: string;
  menu_item_id?: string;
  name?: string;
  preparation_note?: string;
  price_vnd?: number;
  quantity?: number;
  selected_modifier_options?: SalesSelectedModifierOptionResponse[];
  size_id?: string;
  size_name?: string;
}